Move header generator to GUI #51
- fixes #51
- UT :+1:
